+++ grass/models/abstracts/ObjectInfoAbstract.py
