+++ grass/models/abstracts/SimpleResponseAbstract.py
@@ -5,7 +5,7 @@
 # Author: Corey White (smortopahri@gmail.com)                                  #
 # Maintainer: Corey White                                                      #
 # -----                                                                        #
-# Last Modified: Mon Nov 13 2023                                               #
+# Last Modified: Mon Mar 18 2024                                               #
 # Modified By: Corey White                                                     #
 # -----                                                                        #
 # License: GPLv3                                                               #
@@ -33,8 +33,12 @@
 from django.db import models
 
 from grass.models.enums import ResponseStatusEnum
-from .fields.ActiniaResourceStatusEnumField import ActiniaResourceStatusEnumField
-from .fields.ActiniaResponseStatusEnumField import ActiniaResponseStatusEnumField
+from grass.models.fields.ActiniaResourceStatusEnumField import (
+    ActiniaResourceStatusEnumField,
+)
+from grass.models.fields.ActiniaResponseStatusEnumField import (
+    ActiniaResponseStatusEnumField,
+)
 
 
 class SimpleResponseAbstract(models.Model):
